requirements.txt
playwright
PyYAML
langchain-anthropic
langchain-groq
langchain-openai
langchain-ollama
langchain-google
langchain-text-splitters
langchain-community
unstructured[md]
